WIP passing props to Trade
